pyproject.toml
[project]
name = "bbh-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular[sidecar]>=0.29.0",
    "gunicorn>=23.0.0",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
